+++ qadence/backends/pyqtorch/convert_ops.py
@@ -2,38 +2,37 @@ from __future__ import annotations
 
 from functools import reduce
 from itertools import chain as flatten
-from math import log2, prod
+from math import prod
 from operator import add
-from typing import Callable, Sequence, Tuple
+from typing import Sequence, Tuple
 
 import pyqtorch as pyq
 import sympy
 from pyqtorch.apply import apply_operator
 from pyqtorch.matrices import _dagger
-from pyqtorch.parametric import Parametric as PyQParametric
 from pyqtorch.utils import is_diag
 from torch import (
     Tensor,
     argsort,
     bmm,
-    cat,
     cdouble,
-    complex128,
     diag_embed,
     diagonal,
     exp,
     linalg,
-    mean,
-    no_grad,
     ones_like,
     permute,
-    rand,
     tensor,
     transpose,
 )
-from torch import flatten as torchflatten
 from torch.nn import Module
 
+from qadence.backends.utils import (
+    finitediff,
+    infer_batchsize,
+    pyqify,
+    unpyqify,
+)
 from qadence.blocks import (
     AbstractBlock,
     AddBlock,
@@ -68,94 +67,6 @@ supported_gates = list(set(OpName.list()) - set([OpName.TDAGGER]))
 
 Tdagger is currently not supported.
 """
-FINITE_DIFF_EPS = 1e-06
-
-
-def finitediff_sampling(
-    f: Callable, x: Tensor, eps: float = FINITE_DIFF_EPS, num_samples: int = 10
-) -> Tensor:
-    def _finitediff(val: Tensor) -> Tensor:
-        return (f(x + val) - f(x - val)) / (2 * val)  # type: ignore
-
-    with no_grad():
-        return mean(cat([_finitediff(val) for val in rand(1) for _ in range(num_samples)]))
-
-
-def finitediff(f: Callable, x: Tensor, eps: float = FINITE_DIFF_EPS) -> Tensor:
-    return (f(x + eps) - f(x - eps)) / (2 * eps)  # type: ignore
-
-
-def dydx(
-    jacobian: Tensor,
-    qubit_support: tuple,
-    out_state: Tensor,
-    projected_state: Tensor,
-) -> Tensor:
-    return 2 * pyq.overlap(
-        projected_state,
-        apply_operator(
-            state=out_state,
-            operator=jacobian,
-            qubits=qubit_support,
-        ),
-    )
-
-
-def dydxx(
-    op: PyQParametric,
-    values: dict[str, Tensor],
-    out_state: Tensor,
-    projected_state: Tensor,
-) -> Tensor:
-    return 2 * finitediff_sampling(
-        lambda val: dydx(
-            op.jacobian({op.param_name: val}), op.qubit_support, out_state, projected_state
-        ),
-        values[op.param_name],
-    )
-
-
-def pyqify(state: Tensor, n_qubits: int = None) -> Tensor:
-    if n_qubits is None:
-        n_qubits = int(log2(state.shape[1]))
-    if (state.ndim != 2) or (state.size(1) != 2**n_qubits):
-        raise ValueError(
-            "The initial state must be composed of tensors of size "
-            f"(batch_size, 2**n_qubits). Found: {state.size() = }."
-        )
-    return state.T.reshape([2] * n_qubits + [state.size(0)])
-
-
-def unpyqify(state: Tensor) -> Tensor:
-    return torchflatten(state, start_dim=0, end_dim=-2).t()
-
-
-def is_pyq_shape(state: Tensor, n_qubits: int) -> bool:
-    return prod(state.size()[:-1]) == 2**n_qubits  # type: ignore[no-any-return]
-
-
-def is_qadence_shape(state: Tensor, n_qubits: int) -> bool:
-    return state.shape[1] == 2**n_qubits  # type: ignore[no-any-return]
-
-
-def convert_state(state: Tensor, n_qubits: int) -> Tensor:
-    if state.dtype != complex128:
-        raise TypeError(f"Expected type complex128, got {state.dtype}")
-    elif is_qadence_shape(state, n_qubits):
-        return pyqify(state, n_qubits)
-    elif is_pyq_shape(state, n_qubits):
-        return state
-    else:
-        raise ValueError(
-            "Allowed formats for custom initial state are:\
-                  (1) Qadence shape: (batch_size, 2**n_qubits)\
-                  (2) Pyqtorch shape: (2 * n_qubits + [batch_size])\
-                  Found: {state.size() = }"
-        )
-
-
-def infer_batchsize(param_values: dict[str, Tensor] = None) -> int:
-    return max([len(tensor) for tensor in param_values.values()]) if param_values else 1
 
 
 def is_single_qubit_chain(block: AbstractBlock) -> bool:
