+++ qadence/backends/pytorch_wrapper.py
@@ -13,11 +13,7 @@ from torch.nn import Module
 from qadence.backend import Backend as QuantumBackend
 from qadence.backend import Converted, ConvertedCircuit, ConvertedObservable
 from qadence.backends.adjoint import AdjointExpectation
-from qadence.backends.pyqtorch.convert_ops import (
-    convert_state,
-    infer_batchsize,
-)
-from qadence.backends.utils import param_dict
+from qadence.backends.utils import convert_state, infer_batchsize, param_dict
 from qadence.blocks.abstract import AbstractBlock
 from qadence.blocks.primitive import PrimitiveBlock
 from qadence.blocks.utils import uuid_to_block, uuid_to_eigen
